# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Finder API",
        "description": "RESTful API for Book Finder",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/v1/books/import": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Import a book from Google Books only for admin",
                "description": "Import a book into the local database using its Google Books ID. If the book already exists, returns the existing record.",
                "operationId": "1d4eae098997923f5aaf0db4a1ef564b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "google_books_id"
                                ],
                                "properties": {
                                    "google_books_id": {
                                        "description": "Google Books unique identifier",
                                        "type": "string",
                                        "example": "zyTCAlFPjgYC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book already exists in database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book already exists in database"
                                        },
                                        "data": {
                                            "properties": {
                                                "book": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "google_books_id": {
                                                            "type": "string",
                                                            "example": "zyTCAlFPjgYC"
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "The Google Story"
                                                        },
                                                        "author": {
                                                            "type": "string",
                                                            "example": "David A. Vise"
                                                        },
                                                        "isbn": {
                                                            "type": "string",
                                                            "example": "9780553804577"
                                                        },
                                                        "cover_url": {
                                                            "type": "string",
                                                            "example": "http://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Book imported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book imported successfully from Google Books"
                                        },
                                        "data": {
                                            "properties": {
                                                "book": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "google_books_id": {
                                                            "type": "string",
                                                            "example": "zyTCAlFPjgYC"
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "The Google Story"
                                                        },
                                                        "author": {
                                                            "type": "string",
                                                            "example": "David A. Vise"
                                                        },
                                                        "isbn": {
                                                            "type": "string",
                                                            "example": "9780553804577"
                                                        },
                                                        "cover_url": {
                                                            "type": "string",
                                                            "example": "http://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found in Google Books API",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book not found in Google Books API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "google_books_id": [
                                                    "The google books id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to import book",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to import book"
                                        },
                                        "errors": {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "An unexpected error occurred"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "Generic error response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}